requirements.txt
praw==7.6.0
sanic==22.6.0
python-dotenv==0.20.0
